# Conflicts:
#	components/admin_console/admin_definition.jsx

+++ components/admin_console/group_settings/group_row.jsx
@@ -8,10 +8,10 @@ import {FormattedMessage} from 'react-intl';
 
 import {localizeMessage} from 'utils/utils.jsx';
 
-import CheckboxCheckedIcon from 'components/svg/checkbox_checked_icon.jsx';
-import LoadingSpinner from 'components/widgets/loading/loading_spinner.jsx';
+import CheckboxCheckedIcon from 'components/widgets/icons/checkbox_checked_icon.jsx';
+import LoadingSpinner from 'components/widgets/loading/loading_spinner';
 
-export default class GroupRow extends React.Component {
+export default class GroupRow extends React.PureComponent {
     static propTypes = {
         primary_key: PropTypes.string.isRequired,
         name: PropTypes.string.isRequired,
@@ -20,6 +20,7 @@ export default class GroupRow extends React.Component {
         checked: PropTypes.bool,
         failed: PropTypes.bool,
         onCheckToggle: PropTypes.func,
+        readOnly: PropTypes.bool,
         actions: PropTypes.shape({
             link: PropTypes.func.isRequired,
             unlink: PropTypes.func.isRequired,
@@ -34,13 +35,18 @@ export default class GroupRow extends React.Component {
     }
 
     onRowClick = () => {
+        if (this.props.readOnly) {
+            return;
+        }
         this.props.onCheckToggle(this.props.primary_key);
     }
 
     linkHandler = async (e) => {
         e.stopPropagation();
         e.preventDefault();
-
+        if (this.props.readOnly) {
+            return;
+        }
         this.setState({loading: true});
         await this.props.actions.link(this.props.primary_key);
         this.setState({loading: false});
@@ -49,6 +55,9 @@ export default class GroupRow extends React.Component {
     unlinkHandler = async (e) => {
         e.stopPropagation();
         e.preventDefault();
+        if (this.props.readOnly) {
+            return;
+        }
         this.setState({loading: true});
         await this.props.actions.unlink(this.props.primary_key);
         this.setState({loading: false});
@@ -60,7 +69,10 @@ export default class GroupRow extends React.Component {
         }
         if (this.props.has_syncables) {
             return (
-                <Link to={'/admin_console/access-control/groups/' + this.props.mattermost_group_id}>
+                <Link
+                    to={'/admin_console/user_management/groups/' + this.props.mattermost_group_id}
+                    id={`${this.props.name}_edit`}
+                >
                     <FormattedMessage
                         id='admin.group_settings.group_row.edit'
                         defaultMessage='Edit'
@@ -69,7 +81,10 @@ export default class GroupRow extends React.Component {
             );
         }
         return (
-            <Link to={'/admin_console/access-control/groups/' + this.props.mattermost_group_id}>
+            <Link
+                to={'/admin_console/user_management/groups/' + this.props.mattermost_group_id}
+                id={`${this.props.name}_configure`}
+            >
                 <FormattedMessage
                     id='admin.group_settings.group_row.configure'
                     defaultMessage='Configure'
@@ -111,6 +126,7 @@ export default class GroupRow extends React.Component {
                 <a
                     href='#'
                     onClick={this.unlinkHandler}
+                    className={this.props.readOnly ? 'disabled' : ''}
                 >
                     <i className='icon fa fa-link'/>
                     <FormattedMessage
@@ -139,6 +155,7 @@ export default class GroupRow extends React.Component {
             <a
                 href='#'
                 onClick={this.linkHandler}
+                className={this.props.readOnly ? 'disabled' : ''}
             >
                 <i className='icon fa fa-unlink'/>
                 <FormattedMessage
@@ -152,24 +169,29 @@ export default class GroupRow extends React.Component {
     render = () => {
         return (
             <div
+                id={`${this.props.name}_group`}
                 className={'group ' + (this.props.checked ? 'checked' : '')}
                 onClick={this.onRowClick}
             >
                 <div className='group-row'>
-                    <div
-                        className={'group-check ' + (this.props.checked ? 'checked' : '')}
-                    >
-                        {this.props.checked && <CheckboxCheckedIcon/>}
+                    <div className='group-name'>
+                        <div
+                            className={'group-check ' + (this.props.checked ? 'checked' : '')}
+                        >
+                            {this.props.checked && <CheckboxCheckedIcon/>}
+                        </div>
+                        <span>
+                            {this.props.name}
+                        </span>
                     </div>
-                    <span className='group-name'>
-                        {this.props.name}
-                    </span>
-                    <span className='group-description'>
-                        {this.renderLinked()}
-                    </span>
-                    <span className='group-actions'>
-                        {this.renderActions()}
-                    </span>
+                    <div className='group-content'>
+                        <span className='group-description'>
+                            {this.renderLinked()}
+                        </span>
+                        <span className='group-actions'>
+                            {this.renderActions()}
+                        </span>
+                    </div>
                 </div>
             </div>
         );
